added hadler for inactive app
