+++ lib/lock_screen.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class LockScreen extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Container(child: Text("Lock Screen"));
+  }
+}
