+++ lib/main.dart
@@ -1,24 +1,12 @@
-import 'dart:async';
-import 'dart:convert';
-
-import 'package:adv_fab/adv_fab.dart';
-import 'package:fab_circular_menu/fab_circular_menu.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/services.dart';
+import 'package:flutter_samsung_remote/remote.dart';
 import 'package:flutter_samsung_remote/tv_provider.dart';
 import 'package:provider/provider.dart';
-import 'package:rxdart/rxdart.dart';
-import 'package:flutter_vibrate/flutter_vibrate.dart';
-
-import 'app_colors.dart';
 import 'device.dart';
-import 'key_codes.dart';
-import 'package:hardware_buttons/hardware_buttons.dart' as HardwareButtons;
-
-import 'package:virtual_keyboard/virtual_keyboard.dart';
-import 'package:shared_preferences/shared_preferences.dart';
+import 'package:flutter_app_lock/flutter_app_lock.dart';
+import 'package:flutter_samsung_remote/lock_screen.dart';
 
 /// Virtual keyboard actions.
 enum VirtualKeyboardKeyAction { Backspace, Return, Shift, Space }
@@ -29,949 +17,10 @@ void main() async {
   SystemChrome.setPreferredOrientations(
       [DeviceOrientation.portraitUp, DeviceOrientation.portraitDown]);
   SystemChrome.setEnabledSystemUIOverlays([]);
-  return runApp(
-    MultiProvider(
-      providers: [
-        ChangeNotifierProvider(
-          create: (_) => TvProvider(),
-        )
-      ],
-      child: UniversalControllerApp(),
-    ),
-  );
-}
-
-class UniversalControllerApp extends StatelessWidget {
-  final GlobalKey<FabCircularMenuState> fabKey = GlobalKey();
-
-  void toggleFab() {
-    if (fabKey.currentState.isOpen) {
-      fabKey.currentState.close();
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-        debugShowCheckedModeBanner: false,
-        home: Scaffold(
-          body: MyHomePage(),
-          floatingActionButton: FabCircularMenu(
-              key: fabKey,
-              fabElevation: 20,
-              fabSize: 40,
-              fabColor: Colors.cyan[300],
-              fabOpenIcon: Icon(Icons.play_arrow),
-              alignment: Alignment.centerRight,
-              ringColor: AppColors.darkButtonBackground,
-              children: <Widget>[
-                IconButton(
-                    icon: Icon(
-                      Icons.stop,
-                      size: 48,
-                      color: Colors.yellow,
-                    ),
-                    onPressed: () async {
-                      await tv.sendKey(KEY_CODES.KEY_STOP);
-                      toggleFab();
-                    }),
-                IconButton(
-                    icon: Icon(
-                      Icons.fast_rewind_sharp,
-                      size: 28,
-                      color: Colors.white,
-                    ),
-                    onPressed: () async {
-                      await tv.sendKey(KEY_CODES.KEY_REWIND);
-                      toggleFab();
-                    }),
-                IconButton(
-                    icon: Icon(
-                      Icons.play_arrow,
-                      size: 48,
-                      color: Colors.red,
-                    ),
-                    onPressed: () async {
-                      await tv.sendKey(KEY_CODES.KEY_PLAY);
-                      toggleFab();
-                    }),
-                IconButton(
-                    icon: Icon(
-                      Icons.fast_forward,
-                      size: 28,
-                      color: Colors.white,
-                    ),
-                    onPressed: () async {
-                      await tv.sendKey(KEY_CODES.KEY_FF);
-                      toggleFab();
-                    }),
-                IconButton(
-                    icon: Icon(
-                      Icons.pause,
-                      size: 48,
-                      color: Colors.cyanAccent,
-                    ),
-                    onPressed: () async {
-                      await tv.sendKey(KEY_CODES.KEY_PAUSE);
-                      toggleFab();
-                    })
-              ]),
-        ));
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key}) : super(key: key);
-  @override
-  _MyHomePageState createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  BehaviorSubject<int> willAcceptStream;
-  Color backgroundColor = AppColors.darkBackground;
-  Color textColor = AppColors.darkText;
-  Color selectColor = AppColors.darkSelect;
-  Color iconColor = AppColors.darkIcon;
-  Color buttonBackgroundColor = AppColors.darkButtonBackground;
-  Color iconButtonColor = AppColors.darkIconButton;
-  Color sliderBackground = AppColors.darkButtonBackground;
-  String inputValue = "";
-  String _latestHardwareButtonEvent;
-  StreamSubscription<HardwareButtons.VolumeButtonEvent>
-      _volumeButtonSubscription;
-  String token;
-  bool status = false;
-  SharedPreferences _pref;
-  bool _canVibrate;
-  AdvFabController mabialaFABController;
-  bool useFloatingSpaceBar = true;
-  bool useAsFloatingActionButton = false;
-  bool useNavigationBar = false;
-  Map<String, List<String>> myMapList = Map();
-  bool isAlwaysCaps = false;
-  @override
-  void initState() {
-    mabialaFABController = AdvFabController();
-    willAcceptStream = new BehaviorSubject<int>();
-    willAcceptStream.add(0);
-
-    super.initState();
-    init();
-  }
-
-  void init() async {
-    bool canVibrate = await Vibrate.canVibrate;
-    _pref = await SharedPreferences.getInstance();
-    setState(() {
-      _canVibrate = canVibrate;
-      _canVibrate
-          ? print("This device can vibrate")
-          : print("This device cannot vibrate");
-    });
-    _volumeButtonSubscription =
-        HardwareButtons.volumeButtonEvents.listen((event) {
-      setState(() {
-        _latestHardwareButtonEvent = event.toString();
-        volumeButtonActions(_latestHardwareButtonEvent);
-      });
-    });
-    await setUp();
-  }
-
-  Future<void> setUp() async {
-    await wakeTV();
-    await connectTV();
-    if (_pref.getString('token') == null) {
-      await discoverTV();
-      token = await getTvToken();
-      if (tv.token != null) {
-        _pref.setString('token', token);
-        print("Set Token: $token");
-      }
-      var info = await tv.getDeviceInfo();
-      var details = jsonDecode(info.body);
-      _pref.setString('host', tv.host);
-      _pref.setString('mac', details['device']['wifiMac']);
-      print("Set Mac ${details['device']['wifiMac']}");
-    }
-    status = tv.isConnected;
-    setColor(status);
-  }
-
-  Future<void> viewToken() async {
-    print(tv.token);
-  }
-
-  Future<void> wakeTV() async {
-    if (_pref.containsKey('token')) {
-      try {
-        await SamsungSmartTV.wakeOnLan(
-            _pref.getString('host'), _pref.getString('mac'));
-      } catch (e) {
-        print("Failed to Wake on lan");
-      }
-    }
-  }
-
-  Future<void> discoverTV() async {
-    try {
-      tv = await SamsungSmartTV.discover();
-    } catch (e) {
-      print("Failed to discover tv");
-    }
-  }
-
-  Future<void> connectTV() async {
-    token = _pref.getString('token');
-
-    try {
-      if (token != null) {
-        tv = new SamsungSmartTV(
-            deviceName: "Samsung TV",
-            host: _pref.getString('host'),
-            mac: _pref.getString('mac'));
-        await tv.connect(tokenValue: token);
-      }
-    } catch (e) {
-      print("failed to connect");
-    }
-  }
-
-  setColor(status) {
-    Color result;
-    if (status) {
-      result = Colors.green;
-    } else {
-      result = Colors.white;
-    }
-    setState(() {
-      selectColor = result;
-    });
-  }
-
-  Future<String> getTvToken() async {
-    if (_pref.containsKey('token') != true) {
-      await tv.connect(tokenValue: token);
-
-      // token = tv.token;
-
-      // status = tv.isConnected;
-      // setColor(status);
-    }
-
-    return tv.token;
-  }
-
-  void volumeButtonActions(String status) async {
-    switch (status) {
-      case 'VolumeButtonEvent.VOLUME_DOWN':
-        {
-          await tv.sendKey(KEY_CODES.KEY_VOLDOWN);
-        }
-        break;
-
-      case 'VolumeButtonEvent.VOLUME_UP':
-        {
-          await tv.sendKey(KEY_CODES.KEY_VOLUP);
-        }
-        break;
-
-      default:
-        {
-          //statements;
-
-        }
-        break;
-    }
-  }
-
-  void toggleTheme() {
-    setState(
-      () {
-        backgroundColor = backgroundColor == AppColors.darkBackground
-            ? AppColors.lightBackground
-            : AppColors.darkBackground;
-        textColor = textColor == AppColors.darkText
-            ? AppColors.lightText
-            : AppColors.darkText;
-        selectColor = selectColor == AppColors.darkSelect
-            ? AppColors.lightSelect
-            : AppColors.darkSelect;
-        iconColor = iconColor == AppColors.darkIcon
-            ? AppColors.lightIcon
-            : AppColors.darkIcon;
-        iconButtonColor = iconButtonColor == AppColors.darkIconButton
-            ? AppColors.lightIconButton
-            : AppColors.darkIconButton;
-        buttonBackgroundColor =
-            buttonBackgroundColor == AppColors.darkButtonBackground
-                ? AppColors.lightButtonBackground
-                : AppColors.darkButtonBackground;
-      },
-    );
-  }
-
-  void vibrate() {
-    var _type = FeedbackType.selection;
-    Vibrate.feedback(_type);
-  }
-
-  @override
-  void dispose() {
-    super.dispose();
-    _volumeButtonSubscription?.cancel();
-  }
-
-  void setBar() {
-    //3201601007230 HBO
-    //org.tizen.browser internet
-    myMapList['app'] = ["Netflix", "HBO", "Prime", "YouTube"];
-    myMapList['logo'] = [
-      "assets/netflix_logo.jpeg",
-      "assets/hbo_logo.jpeg",
-      "assets/amazon_logo.jpg",
-      "assets/youtube_logo.png"
-    ];
-    mabialaFABController.setExpandedWidgetConfiguration(
-      showLogs: true,
-      heightToExpandTo: 30,
-      expendedBackgroundColor: backgroundColor,
-      withChild: Padding(
-        padding: const EdgeInsets.only(top: 8.0, left: 8.0, right: 8.0),
-        child: Container(
-          width: (MediaQuery.of(context).size.width),
-
-          ///[IMPORTANT]: the height percentage shall be less than [heightToExpandTo]
-          ///in the next line we use 20%
-          height: (MediaQuery.of(context).size.height / 100) * 20,
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: <Widget>[
-              Expanded(
-                child:
-                    Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-                  Padding(
-                    padding: const EdgeInsets.only(top: 8.0, left: 8),
-                    child: Center(
-                        child: GestureDetector(
-                      onTap: () {
-                        mabialaFABController.collapseFAB();
-                      },
-                      child: Text(
-                        'Apps',
-                        style: TextStyle(color: Colors.white),
-                      ),
-                    )),
-                  ),
-                ]),
-              ),
-              Expanded(
-                  flex: 5,
-                  child: ListView.builder(
-                      physics: BouncingScrollPhysics(),
-                      itemCount: 4,
-                      scrollDirection: Axis.horizontal,
-                      itemBuilder: (BuildContext context, int index) {
-                        return Padding(
-                          padding: const EdgeInsets.all(8.0),
-                          child: GestureDetector(
-                            onTap: () {
-                              vibrate();
-                              mabialaFABController.collapseFAB();
-                              tv.openTVApp(myMapList['app'][index]);
-                            },
-                            child: Container(
-                              width: (MediaQuery.of(context).size.width / 100) *
-                                  25,
-                              color: backgroundColor,
-                              child: Image.asset(myMapList['logo'][index]),
-                            ),
-                          ),
-                        );
-                      }))
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    var size = MediaQuery.of(context).size;
-    setBar();
-    return SafeArea(
-        child: Scaffold(
-      floatingActionButton: AdvFab(
-        floatingActionButtonIcon: Icons.apps_outlined,
-        floatingActionButtonIconColor: Colors.red,
-        floatingSpaceBarContainerWidth: 100,
-        useAsFloatingActionButton: true,
-        floatingActionButtonExpendedWidth: 90,
-        useAsNavigationBar: false,
-        controller: mabialaFABController,
-        animationDuration: Duration(milliseconds: 350),
-        useElevation: true,
-        showLogs: false,
-        onFloatingActionButtonTapped: () {
-          vibrate();
-          mabialaFABController.isCollapsed
-              ? mabialaFABController.expandFAB()
-              : mabialaFABController.collapseFAB();
-        },
-      ),
-      floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
-      resizeToAvoidBottomInset: true,
-      body: Container(
-          constraints:
-              BoxConstraints(maxHeight: MediaQuery.of(context).size.height),
-          padding: EdgeInsets.all(10.0),
-          width: size.width,
-          height: size.height,
-          color: backgroundColor,
-          child: FittedBox(
-            fit: BoxFit.cover,
-            alignment: Alignment.center,
-            child: Center(
-              child: Column(
-                  mainAxisAlignment: MainAxisAlignment.spaceAround,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: <Widget>[
-                    Container(
-                      width: size.width,
-                      height: size.height * 0.11,
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceAround,
-                        crossAxisAlignment: CrossAxisAlignment.end,
-                        children: <Widget>[
-                          GestureDetector(
-                            onTap: () async {
-                              vibrate();
-                              print("Mute");
-                              await tv.sendKey(KEY_CODES.KEY_MUTE);
-                            },
-                            child: Container(
-                              width: size.height * 0.11,
-                              height: size.height * 0.08,
-                              child: Icon(
-                                Icons.volume_down,
-                                color: iconColor,
-                                size: 28,
-                              ),
-                            ),
-                          ),
-                          GestureDetector(
-                            onTap: () async {
-                              print("Power Pressed");
-                              vibrate();
-                              await SamsungSmartTV.wakeOnLan(tv.host, tv.mac);
-
-                              await tv.sendKey(KEY_CODES.KEY_POWER);
-                            },
-                            child: Container(
-                              padding: EdgeInsets.all(5),
-                              width: size.height * 0.11,
-                              height: size.height * 0.11,
-                              decoration: new BoxDecoration(
-                                color: buttonBackgroundColor,
-                                shape: BoxShape.circle,
-                              ),
-                              child: Icon(
-                                Icons.power_settings_new,
-                                color: Color(0xFFEF5252),
-                                size: 38,
-                              ),
-                            ),
-                          ),
-                          GestureDetector(
-                            onTap: () {
-                              // toggleTheme();
-                              vibrate();
-                            },
-                            child: Container(
-                              width: size.height * 0.11,
-                              height: size.height * 0.08,
-                              child: Icon(
-                                Icons.personal_video,
-                                color: selectColor,
-                                size: 28,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                    Container(
-                      width: size.width,
-                      height: size.height * 0.25,
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceAround,
-                        children: <Widget>[
-                          CustomButton(
-                            size: size,
-                            buttonBackground: buttonBackgroundColor,
-                            buttonLabel: "Vol",
-                            iconColor: iconButtonColor,
-                            textColor: textColor,
-                            upIcon: Icons.add,
-                            upIconCallBack: () async {
-                              print("Vol up");
-                              vibrate();
-                              await tv.sendKey(KEY_CODES.KEY_VOLUP);
-                            },
-                            downIcon: Icons.remove,
-                            downIconCallBack: () async {
-                              print("Vol down");
-                              vibrate();
-                              await tv.sendKey(KEY_CODES.KEY_VOLDOWN);
-                            },
-                          ),
-                          GestureDetector(
-                            onTap: () async {
-                              await tv.newSendKey("KEY_ENTER");
-                            },
-                            child: CustomCircle(
-                              size: size,
-                              background: backgroundColor,
-                            ),
-                          ),
-                          Container(
-                            width: size.width * 0.20,
-                            height: size.height * 0.25,
-                            decoration: new BoxDecoration(
-                              color: buttonBackgroundColor,
-                              borderRadius: new BorderRadius.all(
-                                Radius.circular(40.0),
-                              ),
-                            ),
-                            child: CustomButton(
-                              size: size,
-                              buttonBackground: buttonBackgroundColor,
-                              buttonLabel: "CH",
-                              iconColor: iconButtonColor,
-                              textColor: textColor,
-                              upIcon: Icons.keyboard_arrow_up,
-                              downIcon: Icons.keyboard_arrow_down,
-                              upIconCallBack: () {
-                                print("Channel Up");
-                                vibrate();
-                              },
-                              downIconCallBack: () {
-                                print("Channel Down");
-                                vibrate();
-                              },
-                            ),
-                          )
-                        ],
-                      ),
-                    ),
-                    Container(
-                      margin: EdgeInsets.only(top: 20),
-                      width: MediaQuery.of(context).size.width,
-                      height: MediaQuery.of(context).size.height / 2.6,
-                      child: Stack(
-                        children: <Widget>[
-                          Positioned(
-                            left: size.width / 2.6,
-                            top: -20,
-                            child: GestureDetector(
-                              onTap: () async {
-                                vibrate();
-                                await tv.sendKey(KEY_CODES.KEY_UP);
-                              },
-                              child: Icon(
-                                Icons.arrow_drop_up,
-                                color: Color(0xFF584BD2),
-                                size: 100,
-                              ),
-                            ),
-                          ),
-                          Positioned(
-                            top: size.height / 8,
-                            left: 20,
-                            child: GestureDetector(
-                              onTap: () async {
-                                vibrate();
-                                await tv.sendKey(KEY_CODES.KEY_LEFT);
-                              },
-                              child: Icon(
-                                Icons.arrow_left,
-                                color: Color(0xFF584BD2),
-                                size: 100,
-                              ),
-                            ),
-                          ),
-                          Positioned(
-                            top: size.height * 0.13,
-                            left: size.width / 2.6,
-                            child: Container(
-                              margin: EdgeInsets.zero,
-                              padding: EdgeInsets.all(5),
-                              width: size.width * 0.26,
-                              height: size.height / 7.5,
-                              decoration: BoxDecoration(
-                                gradient: LinearGradient(
-                                  begin: Alignment.topLeft,
-                                  end: Alignment.bottomRight,
-                                  colors: [
-                                    Colors.transparent,
-                                    Colors.transparent,
-                                    Colors.pinkAccent,
-                                    Colors.blue,
-                                    Color(0xFF584BD2)
-                                  ],
-                                ),
-                                shape: BoxShape.circle,
-                              ),
-                              child: Container(
-                                decoration: new BoxDecoration(
-                                  color: backgroundColor,
-                                  shape: BoxShape.circle,
-                                ),
-                              ),
-                            ),
-                          ),
-                          Positioned(
-                            top: size.height / 8,
-                            right: size.width * 0.025,
-                            child: GestureDetector(
-                              onTap: () async {
-                                vibrate();
-                                await tv.sendKey(KEY_CODES.KEY_RIGHT);
-                              },
-                              child: Icon(
-                                Icons.arrow_right,
-                                color: Color(0xFF584BD2),
-                                size: 100,
-                              ),
-                            ),
-                          ),
-                          Positioned(
-                            top: size.height * 0.13,
-                            left: size.width / 2.6,
-                            child: GestureDetector(
-                              onTap: () async {
-                                vibrate();
-                                await tv.sendKey(KEY_CODES.KEY_ENTER);
-                              },
-                              child: Container(
-                                margin: EdgeInsets.zero,
-                                child: Icon(
-                                  Icons.adjust,
-                                  color: Color(0xFF584BD2),
-                                  size: 100,
-                                ),
-                              ),
-                            ),
-                          ),
-                          Positioned(
-                            bottom: (size.height / 10) - 95,
-                            left: size.width / 2.6,
-                            child: GestureDetector(
-                              onTap: () async {
-                                vibrate();
-                                await tv.sendKey(KEY_CODES.KEY_DOWN);
-                              },
-                              child: Icon(
-                                Icons.arrow_drop_down,
-                                color: Color(0xFF584BD2),
-                                size: 100,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
-                    Container(
-                      width: size.width,
-                      height: size.height / 18,
-                      margin: EdgeInsets.only(top: 20),
-                      child: Column(
-                        mainAxisAlignment: MainAxisAlignment.start,
-                        children: <Widget>[
-                          Row(
-                              mainAxisAlignment: MainAxisAlignment.spaceAround,
-                              children: <Widget>[
-                                GestureDetector(
-                                  onTap: () async {
-                                    vibrate();
-                                    await tv.sendKey(KEY_CODES.KEY_RETURN);
-                                  },
-                                  child: Container(
-                                    child: Icon(
-                                      Icons.arrow_back,
-                                      color: iconColor,
-                                      size: 28,
-                                    ),
-                                  ),
-                                ),
-                                GestureDetector(
-                                  onTap: () async {
-                                    vibrate();
-                                    await tv.sendKey(KEY_CODES.KEY_HOME);
-                                  },
-                                  child: Container(
-                                    child: Icon(
-                                      Icons.home,
-                                      color: iconColor,
-                                      size: 38,
-                                    ),
-                                  ),
-                                ),
-                                GestureDetector(
-                                  onTap: () async {
-                                    print("home");
-                                    await tv.sendKey(KEY_CODES.KEY_MENU);
-                                  },
-                                  child: Icon(
-                                    Icons.settings,
-                                    color: iconColor,
-                                    size: 28,
-                                  ),
-                                ),
-                                GestureDetector(
-                                  onTap: () {
-                                    inputValue = "";
-                                    showModalBottomSheet<void>(
-                                        backgroundColor: Colors.transparent,
-                                        context: context,
-                                        builder: (BuildContext context) {
-                                          return Container(
-                                              margin: EdgeInsets.only(
-                                                  top: 0, bottom: 30),
-                                              height: 400,
-                                              color: Colors.transparent,
-                                              child: Center(
-                                                  child: Container(
-                                                // Keyboard is transparent
-                                                color: Colors.teal,
-                                                child: VirtualKeyboard(
-                                                    alwaysCaps: isAlwaysCaps,
-                                                    // Default height is 300
-                                                    height: 300,
-                                                    // Default is black
-                                                    textColor: Colors.white,
-                                                    // Default 14
-                                                    fontSize: 20,
-                                                    // [A-Z, 0-9]
-                                                    type: VirtualKeyboardType
-                                                        .Alphanumeric,
-                                                    // Callback for key press event
-                                                    onKeyPress: (key) async {
-                                                      if (key.keyType ==
-                                                          VirtualKeyboardKeyType
-                                                              .String) {
-                                                        var txt =
-                                                            (isAlwaysCaps ==
-                                                                    true)
-                                                                ? key.capsText
-                                                                : key.text
-                                                                    .toString();
-                                                        inputValue += txt;
-                                                      } else if (key.keyType ==
-                                                          VirtualKeyboardKeyType
-                                                              .Action) {
-                                                        switch (key.action
-                                                            .toString()) {
-                                                          case "VirtualKeyboardKeyAction.Backspace":
-                                                            print("backspace");
-                                                            if (inputValue
-                                                                    .length ==
-                                                                0) {
-                                                              return;
-                                                            } else {
-                                                              print(
-                                                                  "backspace");
-                                                              inputValue = inputValue
-                                                                  .substring(
-                                                                      0,
-                                                                      inputValue
-                                                                              .length -
-                                                                          1);
-                                                            }
-                                                            break;
-                                                          case "VirtualKeyboardKeyAction.Return":
-                                                            await tv.sendKey(
-                                                                KEY_CODES
-                                                                    .KEY_ENTER);
-                                                            break;
-                                                          case "VirtualKeyboardKeyAction.Space":
-                                                            inputValue += " ";
-                                                            break;
-                                                          case "VirtualKeyboardKeyAction.Shift":
-                                                            setState(() {
-                                                              if (isAlwaysCaps ==
-                                                                  true) {
-                                                                isAlwaysCaps =
-                                                                    false;
-                                                              } else {
-                                                                isAlwaysCaps =
-                                                                    true;
-                                                              }
-                                                            });
-
-                                                            break;
-                                                          default:
-                                                        }
-                                                      }
-                                                      // print(inputValue);
-                                                      await tv.sendInputString(
-                                                          inputValue);
-
-                                                      // await tv.newSendKey(
-                                                      //     "KEY_" + key.text);
-                                                      //search KEY_DTV_SIGNAL
-                                                    }),
-                                              )));
-                                        });
-                                  },
-                                  child: Icon(
-                                    Icons.keyboard,
-                                    color: iconColor,
-                                    size: 28,
-                                  ),
-                                ),
-                              ]),
-                        ],
-                      ),
-                    ),
-                    SizedBox(
-                      width: MediaQuery.of(context).size.width,
-                      height: MediaQuery.of(context).size.height * 0.10,
-                    )
-                  ]),
-            ),
-          )),
-    ));
-  }
-}
-
-class Nub extends StatelessWidget {
-  const Nub({
-    Key key,
-    @required this.size,
-  }) : super(key: key);
-
-  final Size size;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: EdgeInsets.all(3),
-      width: size.width * 0.2,
-      height: size.width * 0.5,
-      child: Icon(
-        Icons.lens,
-        color: Color(0xFFFF4B4D),
-        size: 18,
-      ),
-    );
-  }
-}
-
-class CustomCircle extends StatelessWidget {
-  const CustomCircle({
-    Key key,
-    @required this.size,
-    @required this.background,
-  }) : super(key: key);
-
-  final Size size;
-  final Color background;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      padding: EdgeInsets.all(2),
-      width: size.width * 0.1,
-      height: size.width * 0.1,
-      decoration: new BoxDecoration(
-        gradient: new LinearGradient(
-          begin: Alignment.centerLeft,
-          end: Alignment.centerRight,
-          colors: [
-            Colors.blue,
-            Colors.pink,
-          ],
-        ),
-        shape: BoxShape.circle,
-      ),
-      child: Container(
-        padding: EdgeInsets.all(18),
-        width: size.width * 0.4,
-        height: size.width * 0.4,
-        decoration: new BoxDecoration(
-          color: background,
-          shape: BoxShape.circle,
-        ),
-      ),
-    );
-  }
-}
-
-class CustomButton extends StatelessWidget {
-  const CustomButton(
-      {Key key,
-      @required this.size,
-      @required this.buttonBackground,
-      @required this.iconColor,
-      @required this.buttonLabel,
-      @required this.textColor,
-      @required this.upIcon,
-      @required this.downIcon,
-      @required this.upIconCallBack,
-      @required this.downIconCallBack})
-      : super(key: key);
-
-  final Size size;
-  final Color buttonBackground;
-  final Color iconColor;
-  final Color textColor;
-  final IconData upIcon;
-  final IconData downIcon;
-  final Function upIconCallBack;
-  final Function downIconCallBack;
-  final String buttonLabel;
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: size.width * 0.20,
-      height: size.height * 0.25,
-      decoration: new BoxDecoration(
-        color: buttonBackground,
-        borderRadius: new BorderRadius.all(
-          Radius.circular(40.0),
-        ),
-      ),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceAround,
-        children: <Widget>[
-          GestureDetector(
-            onTap: upIconCallBack,
-            child: Icon(
-              upIcon,
-              color: iconColor,
-              size: 38,
-            ),
-          ),
-          Text(
-            buttonLabel,
-            style: TextStyle(
-              fontWeight: FontWeight.bold,
-              color: textColor,
-              fontSize: 24,
-            ),
-          ),
-          GestureDetector(
-            onTap: downIconCallBack,
-            child: Icon(
-              downIcon,
-              color: iconColor,
-              size: 38,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
+  return runApp(AppLock(
+    builder: (args) => UniversalControllerApp(),
+    lockScreen: LockScreen(),
+    enabled: false,
+    backgroundLockLatency: const Duration(seconds: 15),
+  ));
 }
