+++ pubspec.yaml
@@ -35,6 +35,7 @@ dependencies:
   web_socket_channel:
     git: https://github.com/kecson/web_socket_channel
   upnp: ^2.0.0+1
+  flutter_app_lock: ^1.4.0+1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
